+++ lib/openapi_first/definition/parameters.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'forwardable'
-require_relative 'schema'
+require_relative '../schema'
 
 module OpenapiFirst
   class Parameters