+++ lib/openapi_first/definition/request_body.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'schema'
+require_relative '../schema'
 
 module OpenapiFirst
   class RequestBody